Fixed incorect arguments for stlink_upload
The number of arguments being passed to the stlink_upload script on
linux, linux64, macosx is one (this may have been more in the past but
the with version 1.8.8) this is now only a single argument.

The change changes the variable referenced in the script from $4 to $1.

This change already appears to have been applied to the windows version
of the script.